subtree: Have $indent actually affect indentation
Currently, the $indent variable is just used to track how deeply we're
nested.  Currently, the debug log is indented by things like

   debug "  foo"

That is: The indentation-level is hard-coded; as before 315a84f9aa
("subtree: use commits before rejoins for splits", 2018-09-28) that could
be known statically; it couldn't recurse.  Now it can recurse, and the
debug log is misleading.

So fix that.  Indent according to $indent.