subtree: Parse revs in individual cmd_ functions; avoid globals
The main argument parser goes ahead and tries to parse revs to make things
simpler for the sub-command implementations.  But, it includes enough
special cases for different sub-commands, and having to think about "is
this info coming from an argument, or a global variable?" that it's just
confusing and complicated.

Begone with the 'revs' global variable; parse 'rev=$(...)' as needed in
individual cmd_ functions.

Begone with the 'default' global variable.  Its would-be value is knowable
just from what function we're in.

Begone with the 'ensure_single_rev' function.  Its functionality can be
achieved by passing '--verify' to 'git rev-parse'.