subtree: t7900: Fix and simplify 'verify one file change per commit'
As far as I can tell, this test isn't actually successfully testing,
because someone forgot to tack on `--name-only` to `git log`.  This
seems to have been the case since the test was first written, back in
fa16ab36ad ("test.sh: make sure no commit changes more than one file at
a time.", 2009-04-26), unless `git log` used to do that by default and
didn't need the flag back then?

Convincing myself of this was tricky, the code is a little hard to
reason about.  It can be made a lot simpler if instead of trying to
parse all of the info from a single `git log`, we're OK calling `git
log` from inside of a loop.  And it's my opinion that tests are not the
place for clever optimized code.